+++ src/common/mod.rs
@@ -18,3 +18,4 @@ pub mod telemetry;
 pub mod telemetry_ops;
 pub mod telemetry_reporting;
 pub mod update;
+pub mod validate_vectors;
